Update generalizes objects and create methods.
This replaces the hard coded requirements for Org, Go and txdb
packages and instead stores a data.frame that indicates the
relationships between the packages in question.  This check also
updates cols and keytypes to the new object structure, but select and
keys still need refactoring, and the manual pages and uni tests also
need updates.


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/OrganismDbi@68148 bc3139a8-67e5-0310-9ffc-ced21a209358